Add files via upload
Problem Description :
CredX is a leading credit card provider that gets thousands of credit card applicants every year. But in the past few years, it has experienced an increase in credit loss. The CEO believes that the best strategy to mitigate credit risk is to ‘acquire the right customers’
Objective : 
To help CredX identify the right customers using past data of the bank’s applicants. To determine the factors affecting credit risk, create strategies to mitigate the acquisition risk and assess the financial benefit of the project.